# fabric_ceph/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: |
    Minimal OpenAPI definition for building an app around Ceph Dashboard REST endpoints used to
    manage CephX users and CephFS subvolumes, and to authenticate via JWT. These paths mirror the
    documented `/api` surface (as commonly exposed by the Ceph Dashboard).
  title: Ceph Dashboard – Auth & Cluster User API
  version: 1.0.0
servers:
- description: Local development server
  url: http://localhost:3500
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/RENCI3/ceph/1.0.0
- url: "https://{host}:{port}/api"
  variables:
    host:
      default: mgr.example.org
    port:
      default: "8443"
...
paths:
...
  /s3/user:
    get:
      operationId: list_s3_users
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      - description: Filter by a specific user id (exact match)
        explode: true
        in: query
        name: uid
        required: false
        schema:
          type: string
        style: form
      - description: Case-insensitive substring match on uid or display_name
        explode: true
        in: query
        name: search
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 100
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/S3UserList"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_401_unauthorized"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_403_forbidden"
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_500_internal_server_error"
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List RGW (S3) users
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
    post:
      description: Mirrors radosgw-admin user create/modify with sensible defaults.
      operationId: upsert_s3_user
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrUpdateS3UserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/S3User"
          description: OK (created or updated)
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_400_bad_request"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_401_unauthorized"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_403_forbidden"
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_500_internal_server_error"
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update an RGW user (upsert)
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
  /s3/user/{uid}:
    delete:
      operationId: delete_s3_user
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      - explode: false
        in: path
        name: uid
        required: true
        schema:
          type: string
        style: simple
      - description: Also purge buckets/objects owned by the user (dangerous)
        explode: true
        in: query
        name: purge_data
        required: false
        schema:
          default: false
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_200_ok_no_content"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_404_not_found"
          description: Not Found
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_400_bad_request"
          description: Bad Request
      security:
      - bearerAuth: []
      summary: Delete RGW user
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
    get:
      operationId: get_s3_user
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      - explode: false
        in: path
        name: uid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/S3User"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_404_not_found"
          description: Not Found
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_401_unauthorized"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_403_forbidden"
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_500_internal_server_error"
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get RGW user
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
  /s3/user/{uid}/keys:
    delete:
      operationId: delete_s3_user_key
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      - explode: false
        in: path
        name: uid
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: access_key
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_200_ok_no_content"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_404_not_found"
          description: Not Found
      security:
      - bearerAuth: []
      summary: Delete an access key
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
    get:
      operationId: list_s3_user_keys
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      - explode: false
        in: path
        name: uid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/S3KeyPair"
                type: array
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_404_not_found"
          description: Not Found
      security:
      - bearerAuth: []
      summary: List S3 access keys for a user
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
    post:
      description: "If `access_key` and `secret_key` omitted, a new pair is generated."
      operationId: create_s3_user_key
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      - explode: false
        in: path
        name: uid
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateS3KeyRequest"
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/S3KeyPair"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_404_not_found"
          description: Not Found
      security:
      - bearerAuth: []
      summary: Create or set an access key
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
  /s3/user/{uid}/quota:
    get:
      operationId: get_s3_quota
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      - explode: false
        in: path
        name: uid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/S3Quota"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_404_not_found"
          description: Not Found
      security:
      - bearerAuth: []
      summary: Get user & bucket quota
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
    put:
      operationId: set_s3_quota
      parameters:
      - description: Target cluster/region identifier as defined by the service config.
        explode: true
        in: query
        name: cluster
        required: true
        schema:
          example: europe
          type: string
        style: form
      - explode: false
        in: path
        name: uid
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/S3QuotaRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_200_ok_no_content"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status_400_bad_request"
          description: Bad Request
      security:
      - bearerAuth: []
      summary: Set user and/or bucket quota
      tags:
      - S3 User
      x-openapi-router-controller: fabric_ceph.openapi_server.controllers.s3_user_controller
...
components:
...
  schemas:
...
    status_200_ok_no_content:
      example:
        data:
        - details: details
          message: No Content
        - details: details
          message: No Content
        size: 0
        type: no_content
        status: 6
      properties:
        data:
          items:
            $ref: "#/components/schemas/status_200_ok_no_content_data"
          title: data
          type: array
        type:
          default: no_content
          title: type
          type: string
        size:
          default: 1
          title: size
          type: integer
        status:
          default: 200
          title: status
          type: integer
      title: status_200_ok_no_content
      type: object
    status_200_ok_paginated:
      properties:
        limit:
          title: limit
          type: integer
        offset:
          title: offset
          type: integer
        size:
          title: size
          type: integer
        status:
          default: 200
          title: status
          type: integer
        total:
          title: total
          type: integer
        type:
          title: type
          type: string
      title: status_200_ok_paginated
      type: object
    status_400_bad_request:
      example:
        errors:
        - size: 2
          details: details
          message: Bad Request
          type: error
          status: 7
        - size: 2
          details: details
          message: Bad Request
          type: error
          status: 7
      properties:
        errors:
          items:
            $ref: "#/components/schemas/status_400_bad_request_errors"
          title: errors
          type: array
      title: status_400_bad_request
      type: object
    status_401_unauthorized:
      example:
        size: 9
        type: error
        errors:
        - details: details
          message: Unauthorized
        - details: details
          message: Unauthorized
        status: 3
      properties:
        errors:
          items:
            $ref: "#/components/schemas/status_401_unauthorized_errors"
          title: errors
          type: array
        type:
          default: error
          title: type
          type: string
        size:
          default: 1
          title: size
          type: integer
        status:
          default: 401
          title: status
          type: integer
      title: status_401_unauthorized
      type: object
    status_403_forbidden:
      example:
        size: 2
        type: error
        errors:
        - details: details
          message: Forbidden
        - details: details
          message: Forbidden
        status: 4
      properties:
        errors:
          items:
            $ref: "#/components/schemas/status_403_forbidden_errors"
          title: errors
          type: array
        type:
          default: error
          title: type
          type: string
        size:
          default: 1
          title: size
          type: integer
        status:
          default: 403
          title: status
          type: integer
      title: status_403_forbidden
      type: object
    status_404_not_found:
      example:
        size: 1
        type: error
        errors:
        - details: details
          message: Not Found
        - details: details
          message: Not Found
        status: 5
      properties:
        errors:
          items:
            $ref: "#/components/schemas/status_404_not_found_errors"
          title: errors
          type: array
        type:
          default: error
          title: type
          type: string
        size:
          default: 1
          title: size
          type: integer
        status:
          default: 404
          title: status
          type: integer
      title: status_404_not_found
      type: object
    status_500_internal_server_error:
      example:
        size: 7
        type: error
        errors:
        - details: details
          message: Internal Server Error
        - details: details
          message: Internal Server Error
        status: 1
      properties:
        errors:
          items:
            $ref: "#/components/schemas/status_500_internal_server_error_errors"
          title: errors
          type: array
        type:
          default: error
          title: type
          type: string
        size:
          default: 1
          title: size
          type: integer
        status:
          default: 500
          title: status
          type: integer
      title: status_500_internal_server_error
      type: object
    status_200_ok_no_content_data:
      example:
        details: details
        message: No Content
      properties:
        message:
          default: No Content
          type: string
        details:
          type: object
      title: status_200_ok_no_content_data
      type: object
    status_400_bad_request_errors:
      example:
        size: 2
        details: details
        message: Bad Request
        type: error
        status: 7
      properties:
        message:
          default: Bad Request
          title: message
          type: string
        details:
          title: details
          type: string
        type:
          default: error
          title: type
          type: string
        size:
          default: 1
          title: size
          type: integer
        status:
          default: 400
          title: status
          type: integer
      title: status_400_bad_request_errors
      type: object
    status_401_unauthorized_errors:
      example:
        details: details
        message: Unauthorized
      properties:
        message:
          default: Unauthorized
          title: message
          type: string
        details:
          title: details
          type: string
      title: status_401_unauthorized_errors
      type: object
    status_403_forbidden_errors:
      example:
        details: details
        message: Forbidden
      properties:
        message:
          default: Forbidden
          title: message
          type: string
        details:
          title: details
          type: string
      title: status_403_forbidden_errors
      type: object
    status_404_not_found_errors:
      example:
        details: details
        message: Not Found
      properties:
        message:
          default: Not Found
          title: message
          type: string
        details:
          title: details
          type: string
      title: status_404_not_found_errors
      type: object
    status_500_internal_server_error_errors:
      example:
        details: details
        message: Internal Server Error
      properties:
        message:
          default: Internal Server Error
          title: message
          type: string
        details:
          title: details
          type: string
      title: status_500_internal_server_error_errors
      type: object
...
    S3UserList:
      allOf:
      - $ref: "#/components/schemas/status_200_ok_paginated"
      - properties:
          type:
            example: s3_users
            type: string
          data:
            items:
              $ref: "#/components/schemas/S3User"
            type: array
        type: object
      example:
        total: 5
        offset: 6
        size: 1
        data:
        - uid: demo
          user_quota:
            max_objects: 1000000
            max_size_kb: 104857600
            enabled: false
          system: false
          keys:
          - secret_key: very-secret
            access_key: AKIA...
            user: project123
            status: active
          - secret_key: very-secret
            access_key: AKIA...
            user: project123
            status: active
          max_buckets: 1000
          op_mask: op_mask
          display_name: Demo User
          email: demo@example.org
          suspended: false
          bucket_quota:
            max_objects: 1000000
            max_size_kb: 104857600
            enabled: false
          caps:
          - perm: '*'
            type: buckets
          - perm: '*'
            type: buckets
        - uid: demo
          user_quota:
            max_objects: 1000000
            max_size_kb: 104857600
            enabled: false
          system: false
          keys:
          - secret_key: very-secret
            access_key: AKIA...
            user: project123
            status: active
          - secret_key: very-secret
            access_key: AKIA...
            user: project123
            status: active
          max_buckets: 1000
          op_mask: op_mask
          display_name: Demo User
          email: demo@example.org
          suspended: false
          bucket_quota:
            max_objects: 1000000
            max_size_kb: 104857600
            enabled: false
          caps:
          - perm: '*'
            type: buckets
          - perm: '*'
            type: buckets
        limit: 0
        type: s3_users
        status: 5
      title: S3UserList
    S3User:
      example:
        uid: demo
        user_quota:
          max_objects: 1000000
          max_size_kb: 104857600
          enabled: false
        system: false
        keys:
        - secret_key: very-secret
          access_key: AKIA...
          user: project123
          status: active
        - secret_key: very-secret
          access_key: AKIA...
          user: project123
          status: active
        max_buckets: 1000
        op_mask: op_mask
        display_name: Demo User
        email: demo@example.org
        suspended: false
        bucket_quota:
          max_objects: 1000000
          max_size_kb: 104857600
          enabled: false
        caps:
        - perm: '*'
          type: buckets
        - perm: '*'
          type: buckets
      properties:
        uid:
          example: demo
          title: uid
          type: string
        display_name:
          example: Demo User
          title: display_name
          type: string
        email:
          example: demo@example.org
          nullable: true
          title: email
          type: string
        max_buckets:
          example: 1000
          title: max_buckets
          type: integer
        suspended:
          default: false
          title: suspended
          type: boolean
        caps:
          description: RGW admin caps (if enabled); not IAM
          items:
            $ref: "#/components/schemas/S3User_caps_inner"
          title: caps
          type: array
        keys:
          items:
            $ref: "#/components/schemas/S3KeyPair"
          title: keys
          type: array
        user_quota:
          $ref: "#/components/schemas/S3QuotaEntry"
        bucket_quota:
          $ref: "#/components/schemas/S3QuotaEntry"
        system:
          default: false
          title: system
          type: boolean
        op_mask:
          nullable: true
          title: op_mask
          type: string
      title: S3User
      type: object
    CreateOrUpdateS3UserRequest:
      example:
        uid: project123
        system: false
        keys:
        - secret_key: secret_key
          access_key: access_key
          generate: true
        - secret_key: secret_key
          access_key: access_key
          generate: true
        max_buckets: 0
        op_mask: op_mask
        display_name: Project 123 Service Account
        email: email
        suspended: false
        caps:
        - perm: '*'
          type: users
        - perm: '*'
          type: users
      properties:
        uid:
          example: project123
          title: uid
          type: string
        display_name:
          example: Project 123 Service Account
          title: display_name
          type: string
        email:
          nullable: true
          title: email
          type: string
        suspended:
          default: false
          title: suspended
          type: boolean
        max_buckets:
          default: 1000
          title: max_buckets
          type: integer
        system:
          default: false
          title: system
          type: boolean
        op_mask:
          nullable: true
          title: op_mask
          type: string
        caps:
          items:
            $ref: "#/components/schemas/CreateOrUpdateS3UserRequest_caps_inner"
          title: caps
          type: array
        keys:
          description: Optional explicit keypairs to set; omit to auto-generate none.
          items:
            $ref: "#/components/schemas/CreateS3KeyRequest"
          title: keys
          type: array
      required:
      - uid
      title: CreateOrUpdateS3UserRequest
      type: object
    S3KeyPair:
      example:
        secret_key: very-secret
        access_key: AKIA...
        user: project123
        status: active
      properties:
        user:
          example: project123
          title: user
          type: string
        access_key:
          example: AKIA...
          title: access_key
          type: string
        secret_key:
          example: very-secret
          title: secret_key
          type: string
        status:
          example: active
          title: status
          type: string
      title: S3KeyPair
      type: object
    CreateS3KeyRequest:
      example:
        secret_key: secret_key
        access_key: access_key
        generate: true
      properties:
        access_key:
          nullable: true
          title: access_key
          type: string
        secret_key:
          nullable: true
          title: secret_key
          type: string
        generate:
          default: true
          title: generate
          type: boolean
      title: CreateS3KeyRequest
      type: object
    S3Quota:
      example:
        user_quota:
          max_objects: 1000000
          max_size_kb: 104857600
          enabled: false
        bucket_quota:
          max_objects: 1000000
          max_size_kb: 104857600
          enabled: false
      properties:
        user_quota:
          $ref: "#/components/schemas/S3QuotaEntry"
        bucket_quota:
          $ref: "#/components/schemas/S3QuotaEntry"
      title: S3Quota
      type: object
    S3QuotaEntry:
      example:
        max_objects: 1000000
        max_size_kb: 104857600
        enabled: false
      properties:
        enabled:
          default: false
          title: enabled
          type: boolean
        max_size_kb:
          example: 104857600
          nullable: true
          title: max_size_kb
          type: integer
        max_objects:
          example: 1000000
          nullable: true
          title: max_objects
          type: integer
      title: S3QuotaEntry
      type: object
    S3QuotaRequest:
      example:
        user_quota:
          max_objects: 1000000
          max_size_kb: 104857600
          enabled: false
        bucket_quota:
          max_objects: 1000000
          max_size_kb: 104857600
          enabled: false
      properties:
        user_quota:
          $ref: "#/components/schemas/S3QuotaEntry"
        bucket_quota:
          $ref: "#/components/schemas/S3QuotaEntry"
      title: S3QuotaRequest
      type: object
...
    S3User_caps_inner:
      example:
        perm: '*'
        type: buckets
      properties:
        type:
          example: buckets
          title: type
          type: string
        perm:
          example: '*'
          title: perm
          type: string
      title: S3User_caps_inner
      type: object
    CreateOrUpdateS3UserRequest_caps_inner:
      example:
        perm: '*'
        type: users
      properties:
        type:
          example: users
          title: type
          type: string
        perm:
          example: '*'
          title: perm
          type: string
      title: CreateOrUpdateS3UserRequest_caps_inner
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: fabric_ceph.openapi_server.controllers.security_controller.info_from_bearerAuth